move manifest related procedure to manifestprocessor
